+++ tests/testthat/test-object.R
@@ -18,6 +18,10 @@ test_that("Additional attributes storing properties defined by the class, access
 })
 
 describe("new_object", {
+  it("reports all property type errors", {
+    expect_snapshot(range(start = "x", end = "y"), error = TRUE)
+  })
+
   it("checks new objects for validity", {
     expect_error(range(start = 10, end = 1), "`end` must be greater than or equal to `start`")
   })
@@ -25,13 +29,6 @@ describe("new_object", {
   it("can instantiate a new object that inherits from a basic type", {
     y <- text("foo")
     expect_equal(r7_data(y), as_class("character")("foo"))
-  })
-
-  it("errors if given an invalid property", {
-    expect_error(
-      range(1, "foo"),
-      "must be of class"
-    )
   })
 
   it("checks are arguments are properties", {
@@ -60,7 +57,7 @@ describe("new_object", {
     text2 <- new_class("text2", parent = "character")
     my_class <- new_class("my_class",
       parent = text2,
-      properties = c("name" = "character"),
+      properties = list(name = "character"),
       constructor = function(x, name) new_object(text2(x), name = name)
     )
     obj <- my_class("foo", "bar")