+++ tests/testthat/test-property.R
@@ -36,6 +36,10 @@ describe("prop<-", {
     prop(x, "start") <- 2
     expect_equal(prop(x, "start"), 2)
   })
+  it("errors if the property doesn't exist", {
+    x <- range(1, 10)
+    expect_snapshot(error = TRUE, x@foo <- 10)
+  })
   it("errors if the value does not match the correct class", {
     x <- range(1, 10)
     expect_error(
@@ -47,6 +51,15 @@ describe("prop<-", {
     x <- range(1, 10)
     prop(x, "end", check = FALSE) <- "foo"
     expect_equal(x@end, "foo")
+  })
+})
+
+describe("props<-", {
+  it("validates after setting all properties", {
+    x <- range(1, 2)
+    props(x) <- list(start = 5, end = 10)
+    expect_equal(x@start, 5)
+    expect_equal(x@end, 10)
   })
 })
 
@@ -147,6 +160,14 @@ test_that("property setters can set themselves", {
   expect_equal(x@bar, "foo-bar")
 })
 
+
+test_that("new_property validates name", {
+  expect_snapshot(error = TRUE, {
+    new_property(1)
+    new_property("")
+  })
+})
+
 test_that("properties can be base, S3, S4, R7, or R7 union", {
   class_r7 <- new_class("class_r7")
   class_s4 <- methods::setClass("class_s4", slots = c(x = "numeric"))
@@ -186,5 +207,26 @@ test_that("properties can be base, S3, S4, R7, or R7 union", {
     my_obj@s4 <- "x"
     my_obj@r7 <- "x"
     my_obj@r7_union <- "x"
+  })
+})
+
+test_that("as_properties normalises properties", {
+  expect_equal(as_properties(NULL), list())
+  expect_equal(
+    as_properties(list(new_property("y"))),
+    list(y = new_property("y")
+  ))
+  expect_equal(
+    as_properties(list(x = "numeric")),
+    list(x = new_property("x", "numeric")
+  ))
+})
+
+test_that("as_properties() gives useful error messages", {
+  expect_snapshot(error = TRUE, {
+    as_properties(1)
+    as_properties(list(1))
+    as_properties(list(x = 1))
+    as_properties(list(x = "character", x = "character"))
   })
 })